requirements.txt
Django==5.1.2
django-cors-headers==4.3.1
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-standardized-errors==0.12.6
factory-boy==3.3.0
Faker==26.0.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
pytest
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-env==1.1.3
pytest-mock==3.14.0
whitenoise==6.8.2
